// Angular-pi/light-bootstrap-dashboard-angular2-master/src/app/models/user.ts
export class user {
 idUser: any;
    firstName: any;
    lastName: any;
     username: any;
     age: any;
    birthDate: any;
     placeBirth: any;
     job: any;
     sex: any;
     housing: any;
     postalCode: any;
     email: any;
     actived: any;
    numPhone: any;
     password: any;
     dateCreation: any;
     role : any;
     banned: any;
     bannedPeriode:any;
     salary:any;
     
}